+++ instrumentation/ethon/lib/opentelemetry/instrumentation/ethon/patches/stable/easy.rb
@@ -12,20 +12,11 @@ module OpenTelemetry
         module Stable
           # Ethon::Easy patch for instrumentation
           module Easy
-            ACTION_NAMES_TO_HTTP_METHODS = Hash.new do |h, k|
-              # #to_s is required because user input could be symbol or string
-              h[k] = k.to_s.upcase
-            end
-            HTTP_METHODS_TO_SPAN_NAMES = Hash.new do |h, k|
-              h[k] = k.to_s
-              h[k] = 'HTTP' if k == '_OTHER'
-            end
-
             # Constant for the HTTP status range
             HTTP_STATUS_SUCCESS_RANGE = (100..399)
 
             def http_request(url, action_name, options = {})
-              @otel_method = ACTION_NAMES_TO_HTTP_METHODS[action_name]
+              @otel_method = action_name
               super
             end
 
@@ -76,12 +67,11 @@ module OpenTelemetry
             end
 
             def otel_before_request
-              method = '_OTHER' # Could be GET or not HTTP at all
-              method = @otel_method if instance_variable_defined?(:@otel_method) && !@otel_method.nil?
+              span_data = HttpHelper.span_attrs_for(@otel_method)
 
               @otel_span = tracer.start_span(
-                HTTP_METHODS_TO_SPAN_NAMES[method],
-                attributes: span_creation_attributes(method),
+                span_data.span_name,
+                attributes: span_creation_attributes(span_data),
                 kind: :client
               )
 
@@ -98,10 +88,11 @@ module OpenTelemetry
 
             private
 
-            def span_creation_attributes(method)
+            def span_creation_attributes(span_data)
               instrumentation_attrs = {
-                'http.request.method' => method
+                'http.request.method' => span_data.normalized_method
               }
+              instrumentation_attrs['http.request.method_original'] = span_data.original_method if span_data.original_method
 
               uri = _otel_cleanse_uri(url)
               if uri