+++ instrumentation/http_client/test/instrumentation/http_client/patches/dup/client_test.rb
@@ -106,7 +106,7 @@ describe OpenTelemetry::Instrumentation::HttpClient::Patches::Dup::Client do
       _(span.attributes['http.target']).must_equal '/timeout'
       _(span.attributes['net.peer.name']).must_equal 'example.com'
       _(span.attributes['net.peer.port']).must_equal 443
-      # stable semantic coventions
+      # stable semantic conventions
       _(span.attributes['http.request.method']).must_equal 'GET'
       _(span.attributes['url.scheme']).must_equal 'https'
       _(span.attributes['http.response.status_code']).must_be_nil
@@ -169,6 +169,32 @@ describe OpenTelemetry::Instrumentation::HttpClient::Patches::Dup::Client do
       assert_requested(
         :get,
         'http://example.com/success?hello=there',
+        headers: { 'Traceparent' => "00-#{span.hex_trace_id}-#{span.hex_span_id}-01" }
+      )
+    end
+
+    it 'traces a request with non-standard HTTP method' do
+      stub_request(:purge, 'http://example.com/cache').to_return(status: 200)
+      http = HTTPClient.new
+      http.request(:purge, 'http://example.com/cache')
+
+      _(span.name).must_equal 'HTTP'
+      # old semantic conventions
+      _(span.attributes['http.method']).must_equal '_OTHER'
+      _(span.attributes['http.status_code']).must_equal 200
+      _(span.attributes['http.scheme']).must_equal 'http'
+      _(span.attributes['net.peer.name']).must_equal 'example.com'
+      _(span.attributes['http.target']).must_equal '/cache'
+      # stable semantic conventions
+      _(span.attributes['http.request.method']).must_equal '_OTHER'
+      _(span.attributes['http.request.method_original']).must_equal 'PURGE'
+      _(span.attributes['http.response.status_code']).must_equal 200
+      _(span.attributes['url.scheme']).must_equal 'http'
+      _(span.attributes['server.address']).must_equal 'example.com'
+      _(span.attributes['url.path']).must_equal '/cache'
+      assert_requested(
+        :purge,
+        'http://example.com/cache',
         headers: { 'Traceparent' => "00-#{span.hex_trace_id}-#{span.hex_span_id}-01" }
       )
     end